requirements.txt
discord.py==2.6.3
aiosqlite==0.21.0
tzdata